+++ CHANGES.rst
@@ -2,6 +2,13 @@
 Changelog
 =========
 
+1.1.0 - develop
+===============
+
+- ``route`` and ``command`` may be used directly, not only as decorator
+- add support class-based views for ``route`` decorator
+- add ``get_app_label`` shortcut
+
 1.0.0 - 2016-09-30
 ==================